renaming properypage to propertydetailspage for clarity
