+++ mini-cap/src/pages/PropertyDetailsPage.jsx
